Move network-related classes to a separate package
