+++ app/src/main/java/fr/gaulupeau/apps/Poche/network/tasks/DeleteArticleTask.java
@@ -1,4 +1,4 @@
-package fr.gaulupeau.apps.Poche.data;
+package fr.gaulupeau.apps.Poche.network.tasks;
 
 import android.content.Context;
 import android.widget.Toast;
@@ -10,9 +10,9 @@ import fr.gaulupeau.apps.Poche.entity.Article;
 import fr.gaulupeau.apps.Poche.entity.ArticleDao;
 import fr.gaulupeau.apps.Poche.ui.ConnectionFailAlert;
 
-public class ToggleArchiveTask extends GenericArticleTask {
+public class DeleteArticleTask extends GenericArticleTask {
 
-    public ToggleArchiveTask(Context context, int articleId, ArticleDao articleDao, Article article) {
+    public DeleteArticleTask(Context context, int articleId, ArticleDao articleDao, Article article) {
         super(context, articleId, articleDao, article);
     }
 
@@ -20,17 +20,16 @@ public class ToggleArchiveTask extends GenericArticleTask {
     protected void onPreExecute() {
         super.onPreExecute();
 
-        article.setArchive(!article.getArchive());
-        articleDao.update(article);
+        articleDao.delete(article);
     }
 
     @Override
     protected Boolean doInBackgroundSimple(Void... params) throws IOException {
         if(isOffline) return false;
 
-        if(service.toggleArchive(articleId)) return true;
+        if(service.deleteArticle(articleId)) return true;
 
-        if(context != null) errorMessage = context.getString(R.string.toggleArchive_errorMessage);
+        if(context != null) errorMessage = context.getString(R.string.deleteArticle_errorMessage);
         return false;
     }
 
@@ -38,18 +37,12 @@ public class ToggleArchiveTask extends GenericArticleTask {
     protected void onPostExecute(Boolean success) {
         super.onPostExecute(success);
 
-        article.setSync(success); // ?
-        articleDao.update(article);
-
         if(success || isOffline) {
             if(context != null) {
-                Toast.makeText(context, article.getArchive()
-                                ? R.string.moved_to_archive_message
-                                : R.string.marked_as_unread_message,
-                        Toast.LENGTH_SHORT).show();
+                Toast.makeText(context, R.string.deleteArticle_deleted, Toast.LENGTH_SHORT).show();
 
                 if(isOffline) {
-                    Toast.makeText(context, R.string.toggleArchive_noInternetConnection,
+                    Toast.makeText(context, R.string.deleteArticle_noInternetConnection,
                             Toast.LENGTH_SHORT).show();
                 }
             }
