+++ app/src/main/java/fr/gaulupeau/apps/Poche/network/tasks/FeedUpdaterInterface.java
@@ -1,4 +1,4 @@
-package fr.gaulupeau.apps.Poche.data;
+package fr.gaulupeau.apps.Poche.network.tasks;
 
 /**
  * Created by kevinmeyer on 13/12/14.
