+++ app/src/main/java/fr/gaulupeau/apps/Poche/network/tasks/ToggleArchiveTask.java
@@ -1,4 +1,4 @@
-package fr.gaulupeau.apps.Poche.data;
+package fr.gaulupeau.apps.Poche.network.tasks;
 
 import android.content.Context;
 import android.widget.Toast;
@@ -10,9 +10,9 @@ import fr.gaulupeau.apps.Poche.entity.Article;
 import fr.gaulupeau.apps.Poche.entity.ArticleDao;
 import fr.gaulupeau.apps.Poche.ui.ConnectionFailAlert;
 
-public class ToggleFavoriteTask extends GenericArticleTask {
+public class ToggleArchiveTask extends GenericArticleTask {
 
-    public ToggleFavoriteTask(Context context, int articleId, ArticleDao articleDao, Article article) {
+    public ToggleArchiveTask(Context context, int articleId, ArticleDao articleDao, Article article) {
         super(context, articleId, articleDao, article);
     }
 
@@ -20,7 +20,7 @@ public class ToggleFavoriteTask extends GenericArticleTask {
     protected void onPreExecute() {
         super.onPreExecute();
 
-        article.setFavorite(!article.getFavorite());
+        article.setArchive(!article.getArchive());
         articleDao.update(article);
     }
 
@@ -28,9 +28,9 @@ public class ToggleFavoriteTask extends GenericArticleTask {
     protected Boolean doInBackgroundSimple(Void... params) throws IOException {
         if(isOffline) return false;
 
-        if(service.toggleFavorite(articleId)) return true;
+        if(service.toggleArchive(articleId)) return true;
 
-        if(context != null) errorMessage = context.getString(R.string.toggleFavorite_errorMessage);
+        if(context != null) errorMessage = context.getString(R.string.toggleArchive_errorMessage);
         return false;
     }
 
@@ -43,13 +43,13 @@ public class ToggleFavoriteTask extends GenericArticleTask {
 
         if(success || isOffline) {
             if(context != null) {
-                Toast.makeText(context, article.getFavorite()
-                                ? R.string.added_to_favorites_message
-                                : R.string.removed_from_favorites_message,
+                Toast.makeText(context, article.getArchive()
+                                ? R.string.moved_to_archive_message
+                                : R.string.marked_as_unread_message,
                         Toast.LENGTH_SHORT).show();
 
                 if(isOffline) {
-                    Toast.makeText(context, R.string.toggleFavorite_noInternetConnection,
+                    Toast.makeText(context, R.string.toggleArchive_noInternetConnection,
                             Toast.LENGTH_SHORT).show();
                 }
             }
