+++ app/src/main/java/fr/gaulupeau/apps/Poche/network/tasks/ToggleFavoriteTask.java
@@ -1,4 +1,4 @@
-package fr.gaulupeau.apps.Poche.data;
+package fr.gaulupeau.apps.Poche.network.tasks;
 
 import android.content.Context;
 import android.widget.Toast;
@@ -10,9 +10,9 @@ import fr.gaulupeau.apps.Poche.entity.Article;
 import fr.gaulupeau.apps.Poche.entity.ArticleDao;
 import fr.gaulupeau.apps.Poche.ui.ConnectionFailAlert;
 
-public class DeleteArticleTask extends GenericArticleTask {
+public class ToggleFavoriteTask extends GenericArticleTask {
 
-    public DeleteArticleTask(Context context, int articleId, ArticleDao articleDao, Article article) {
+    public ToggleFavoriteTask(Context context, int articleId, ArticleDao articleDao, Article article) {
         super(context, articleId, articleDao, article);
     }
 
@@ -20,16 +20,17 @@ public class DeleteArticleTask extends GenericArticleTask {
     protected void onPreExecute() {
         super.onPreExecute();
 
-        articleDao.delete(article);
+        article.setFavorite(!article.getFavorite());
+        articleDao.update(article);
     }
 
     @Override
     protected Boolean doInBackgroundSimple(Void... params) throws IOException {
         if(isOffline) return false;
 
-        if(service.deleteArticle(articleId)) return true;
+        if(service.toggleFavorite(articleId)) return true;
 
-        if(context != null) errorMessage = context.getString(R.string.deleteArticle_errorMessage);
+        if(context != null) errorMessage = context.getString(R.string.toggleFavorite_errorMessage);
         return false;
     }
 
@@ -37,12 +38,18 @@ public class DeleteArticleTask extends GenericArticleTask {
     protected void onPostExecute(Boolean success) {
         super.onPostExecute(success);
 
+        article.setSync(success); // ?
+        articleDao.update(article);
+
         if(success || isOffline) {
             if(context != null) {
-                Toast.makeText(context, R.string.deleteArticle_deleted, Toast.LENGTH_SHORT).show();
+                Toast.makeText(context, article.getFavorite()
+                                ? R.string.added_to_favorites_message
+                                : R.string.removed_from_favorites_message,
+                        Toast.LENGTH_SHORT).show();
 
                 if(isOffline) {
-                    Toast.makeText(context, R.string.deleteArticle_noInternetConnection,
+                    Toast.makeText(context, R.string.toggleFavorite_noInternetConnection,
                             Toast.LENGTH_SHORT).show();
                 }
             }
