+++ app/src/main/java/fr/gaulupeau/apps/Poche/ui/AddActivity.java
@@ -7,7 +7,7 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 
 import fr.gaulupeau.apps.InThePoche.R;
-import fr.gaulupeau.apps.Poche.data.AddLinkTask;
+import fr.gaulupeau.apps.Poche.network.tasks.AddLinkTask;
 
 public class AddActivity extends AppCompatActivity {
