+++ app/src/main/java/fr/gaulupeau/apps/Poche/ui/ListArticlesActivity.java
@@ -24,12 +24,12 @@ import de.greenrobot.dao.query.QueryBuilder;
 import fr.gaulupeau.apps.InThePoche.BuildConfig;
 import fr.gaulupeau.apps.InThePoche.R;
 import fr.gaulupeau.apps.Poche.data.DbConnection;
-import fr.gaulupeau.apps.Poche.data.FeedUpdater;
-import fr.gaulupeau.apps.Poche.data.FeedUpdaterInterface;
+import fr.gaulupeau.apps.Poche.network.tasks.FeedUpdater;
+import fr.gaulupeau.apps.Poche.network.tasks.FeedUpdaterInterface;
 import fr.gaulupeau.apps.Poche.data.ListAdapter;
 import fr.gaulupeau.apps.Poche.data.Settings;
-import fr.gaulupeau.apps.Poche.data.UploadOfflineURLsTask;
-import fr.gaulupeau.apps.Poche.data.WallabagConnection;
+import fr.gaulupeau.apps.Poche.network.tasks.UploadOfflineURLsTask;
+import fr.gaulupeau.apps.Poche.network.WallabagConnection;
 import fr.gaulupeau.apps.Poche.data.WallabagSettings;
 import fr.gaulupeau.apps.Poche.entity.Article;
 import fr.gaulupeau.apps.Poche.entity.ArticleDao;
