+++ app/src/main/java/fr/gaulupeau/apps/Poche/ui/PocheActivity.java
@@ -24,8 +24,8 @@ import de.greenrobot.dao.query.LazyList;
 import fr.gaulupeau.apps.InThePoche.BuildConfig;
 import fr.gaulupeau.apps.InThePoche.R;
 import fr.gaulupeau.apps.Poche.App;
-import fr.gaulupeau.apps.Poche.data.FeedUpdater;
-import fr.gaulupeau.apps.Poche.data.FeedUpdaterInterface;
+import fr.gaulupeau.apps.Poche.network.tasks.FeedUpdater;
+import fr.gaulupeau.apps.Poche.network.tasks.FeedUpdaterInterface;
 import fr.gaulupeau.apps.Poche.data.DbConnection;
 import fr.gaulupeau.apps.Poche.data.Settings;
 import fr.gaulupeau.apps.Poche.entity.ArticleDao;
